+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+- JSON output for plans and batch summaries.
+
 ## [0.1.5] - 2025-12-22
 
 ### Added